+++ src/main/java/ru/yandex/practicum/filmorate/controller/UserController.java
@@ -3,6 +3,7 @@ package ru.yandex.practicum.filmorate.controller;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
+import ru.yandex.practicum.filmorate.exceptions.ValidationException;
 import ru.yandex.practicum.filmorate.model.User;
 import ru.yandex.practicum.filmorate.service.UserService;
 import ru.yandex.practicum.filmorate.storage.user.UserStorage;
@@ -51,12 +52,20 @@ public class UserController {
 
     @PutMapping("/{userId}/friends/{friendId}")
     public User addFriend(@PathVariable Long userId, @PathVariable Long friendId) {
+        if(userId == friendId) {
+            log.error("UserID and FriendID should be different");
+            throw new ValidationException("UserID and FriendID should be different");
+        }
         log.info("Making friends id {} and {}", userId, friendId);
         return userService.addFriend(userId, friendId);
     }
 
     @DeleteMapping("/{userId}/friends/{friendId}")
     public User deleteFriend(@PathVariable Long userId, @PathVariable Long friendId) {
+        if(userId == friendId) {
+            log.error("UserID and FriendID should be different");
+            throw new ValidationException("UserID and FriendID should be different");
+        }
         log.info("Deleting friends id {} and {}", userId, friendId);
         return userService.deleteFriend(userId, friendId);
     }
@@ -68,7 +77,7 @@ public class UserController {
     }
 
     @GetMapping("/{id}/friends/common/{otherId}")
-    public List<User> getCommonFriends(@PathVariable Long id, @PathVariable Long otherId) {
+    public List<User> getCommonFriends(@PathVariable() Long id, @PathVariable Long otherId) {
         log.info("Looking for common friends for {}  and {}", id, otherId);
         return userService.getCommonFriends(id, otherId);
     }
